+++ site/views/date-picker/base-demo.vue
@@ -3,6 +3,5 @@
 </template>
 
 <script setup>
-// 变量
-const dateValue = '2021-12-23';
+const dateValue = new Date();
 </script>
