+++ site/views/date-picker/range-demo.vue
@@ -1,0 +1,11 @@
+<template>
+  <bk-date-picker
+    class="mr15"
+    :placeholder="'选择日期范围'"
+    :type="'daterange'"
+  />
+</template>
+
+<script setup>
+// const initValArr = [new Date(), new Date()];
+</script>
